Started working on login and registration page

+++ Models/User.cs
@@ -13,6 +13,6 @@ namespace WPF_Dusza.Models
         public string Password { get; set; }
         public int Points { get; set; }
         // 0 - admin, 1 - szervező, 2 - játékos
-        public byte Role { get; set; }
+        public int Role { get; set; }
     }
 }
